IOS-5346 Rename secp256k1 header
