+++ TangemSdk/TangemSdk/Crypto/secp256k1/include/secp256k1_schnorrsig.h
@@ -1,7 +1,7 @@
 #ifndef SECP256K1_SCHNORRSIG_H
 #define SECP256K1_SCHNORRSIG_H
 
-#include "secp256k1.h"
+#include "secp256k1_tangem.h"
 #include "secp256k1_extrakeys.h"
 
 #ifdef __cplusplus
