+++ TangemSdk/TangemSdk/Crypto/secp256k1/src/precomputed_ecmult.c
@@ -2,7 +2,7 @@
 /* This file contains an array secp256k1_pre_g with odd multiples of the base point G and
  * an array secp256k1_pre_g_128 with odd multiples of 2^128*G for accelerating the computation of a*P + b*G.
  */
-#include "../include/secp256k1.h"
+#include "../include/secp256k1_tangem.h"
 #include "group.h"
 #include "ecmult.h"
 #include "precomputed_ecmult.h"
