+++ TangemSdk/TangemSdk/Crypto/secp256k1/src/precomputed_ecmult_gen.c
@@ -1,6 +1,6 @@
 /* This file was automatically generated by precompute_ecmult_gen. */
 /* See ecmult_gen_impl.h for details about the contents of this file. */
-#include "../include/secp256k1.h"
+#include "../include/secp256k1_tangem.h"
 #include "group.h"
 #include "ecmult_gen.h"
 #include "precomputed_ecmult_gen.h"
