+++ TangemSdk/TangemSdk/Crypto/secp256k1/src/util.h
@@ -7,7 +7,7 @@
 #ifndef SECP256K1_UTIL_H
 #define SECP256K1_UTIL_H
 
-#include "../include/secp256k1.h"
+#include "../include/secp256k1_tangem.h"
 
 #include <stdlib.h>
 #include <stdint.h>
