+++ TangemSdk/TangemSdk/module.modulemap
@@ -4,6 +4,6 @@ module TangemSdk_secp256k1 [system] {
   header "Crypto/secp256k1/include/secp256k1_preallocated.h"
   header "Crypto/secp256k1/include/secp256k1_schnorrsig.h"
   header "Crypto/secp256k1/include/secp256k1_extrakeys.h"
-  header "Crypto/secp256k1/include/secp256k1.h"
+  header "Crypto/secp256k1/include/secp256k1_tangem.h" //secp256k1.h leads to conflicts with another secp256k1 copies when installing locally
   export *
 }
